move most working logic to ExpressionPrepender; NodeVisitor should be only delegator
