+++ src/NodeVisitor/ExpressionPrependingNodeVisitor.php
@@ -3,10 +3,7 @@
 namespace Rector\NodeVisitor;
 
 use PhpParser\Node;
-use PhpParser\Node\Expr;
-use PhpParser\Node\Stmt\Expression;
 use PhpParser\NodeVisitorAbstract;
-use Rector\NodeTraverserQueue\BetterNodeFinder;
 use SplObjectStorage;
 
 final class ExpressionPrependingNodeVisitor extends NodeVisitorAbstract
@@ -15,21 +12,6 @@ final class ExpressionPrependingNodeVisitor extends NodeVisitorAbstract
      * @var SplObjectStorage
      */
     private $expressionsToPrepend;
-
-    /**
-     * @var BetterNodeFinder
-     */
-    private $betterNodeFinder;
-
-    public function __construct(BetterNodeFinder $betterNodeFinder)
-    {
-        $this->betterNodeFinder = $betterNodeFinder;
-    }
-
-    public function clear(): void
-    {
-        $this->expressionsToPrepend = new SplObjectStorage();
-    }
 
     public function setExpressionsToPrepend(SplObjectStorage $expressionsToPrepend)
     {
@@ -46,25 +28,5 @@ final class ExpressionPrependingNodeVisitor extends NodeVisitorAbstract
         }
 
         return array_merge([$node], $this->expressionsToPrepend[$node]);
-    }
-
-    public function prependNodeAfterNode(Expr $nodeToPrepend, Node $positionNode): void
-    {
-        $expressionToPrepend = $this->wrapToExpression($nodeToPrepend);
-        $positionExpressionNode = $this->betterNodeFinder->findFirstAncestorInstanceOf($positionNode, Expression::class);
-
-        if (isset($this->expressionsToPrepend[$positionExpressionNode])) {
-            $this->expressionsToPrepend[$positionExpressionNode] = array_merge(
-                $this->expressionsToPrepend[$positionExpressionNode],
-                [$expressionToPrepend]
-            );
-        } else {
-            $this->expressionsToPrepend[$positionExpressionNode] = [$expressionToPrepend];
-        }
-    }
-
-    private function wrapToExpression(Expr $exprNode): Expression
-    {
-        return new Expression($exprNode);
     }
 }
